% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActivateSmoothR.R
\name{ActivateRunSmoothR}
\alias{ActivateRunSmoothR}
\title{Activate RunSmoothR.sh by Copying and Adding to $PATH}
\usage{
ActivateRunSmoothR(dest_folder)
}
\arguments{
\item{dest_folder}{The directory to which the script will be copied and added to $PATH.}
}
\description{
This function copies 'RunSmoothR.sh' to a specified directory and adds the directory
to the $PATH variable by updating the appropriate shell configuration file,
only if the path is not already present.
}
