% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixEQTLwrapper.R, R/matrixEQTLwrapperMC.R
\name{check_complete_match}
\alias{check_complete_match}
\title{Check Complete Matches for All Elements}
\usage{
check_complete_match(sample_name_vector, group_name_vector)

check_complete_match(sample_name_vector, group_name_vector)
}
\arguments{
\item{sample_name_vector}{Character vector; strings to match against group_name_vector.}

\item{group_name_vector}{Character vector; vector within which to search for each element of sample_name_vector.}
}
\value{
Logical; TRUE if all elements from sample_name_vector match any element in group_name_vector, otherwise FALSE.

Logical; TRUE if all elements from sample_name_vector match any element in group_name_vector, otherwise FALSE.
}
\description{
This function verifies whether all elements of sample_name_vector find a match
in group_name_vector using the check_match function. It returns TRUE if all elements match,
otherwise FALSE.

This function verifies whether all elements of sample_name_vector find a match
in group_name_vector using the check_match function. It returns TRUE if all elements match,
otherwise FALSE.
}
\keyword{internal}
