% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixEQTLwrapper.R
\name{check_match}
\alias{check_match}
\title{Check for Partial Matches in Vector}
\usage{
check_match(sample_name, group_name_vector)
}
\arguments{
\item{sample_name}{Character; a single string to look for within group_name_vector.}

\item{group_name_vector}{Character vector; vector within which to search for sample_name.}
}
\value{
Logical; TRUE if there is any match, otherwise FALSE.
}
\description{
This function checks if any part of a given sample name appears in any of the elements
of a vector of group names. It returns TRUE if any match is found, otherwise FALSE.
}
\keyword{internal}
