% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{checkDir}
\alias{checkDir}
\title{Change Working Directory or Verify Current Directory}
\usage{
checkDir(output_dir)
}
\description{
This function checks if the specified `output_dir` exists as a directory path.
If it exists, the function changes the current working directory to `output_dir`.
If it does not exist, the function checks if the current directory's name matches `output_dir`.
If the current directory matches, it confirms the location; otherwise, it throws an error.
}
