% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpoint.R
\name{checkpoint}
\alias{checkpoint}
\title{Prints a Checkpoint Message with a Custom Comment in SLURM Output}
\usage{
checkpoint(comment)
}
\arguments{
\item{comment}{A character string representing the comment to be appended to the checkpoint message.}
}
\value{
Prints a formatted checkpoint message to the console; returns invisible NULL.
}
\description{
This function outputs a formatted checkpoint message that includes a user-defined comment.
}
\examples{
checkpoint("Data loaded successfully")
checkpoint("Model training started")
}
