% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActivateSmoothR.R
\name{CopyRunSmoothRScript}
\alias{CopyRunSmoothRScript}
\title{Copy runSmoothR.sh Script to a Specified Directory}
\usage{
CopyRunSmoothRScript(dest_folder)
}
\arguments{
\item{dest_folder}{The destination folder where the script will be copied.}
}
\description{
This function copies the 'runSmoothR.sh' script from the 'shell_helpers' folder
of the package to a specified folder.
}
