% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_script_name.R
\name{get_script_name}
\alias{get_script_name}
\title{Retrieve the Name of the Currently Running R Script}
\usage{
get_script_name()
}
\value{
A character string containing the filename of the R script being executed,
        or NULL if the function cannot determine the script name (e.g., when run interactively).
}
\description{
This function extracts the name of the R script that is currently being executed.
It is intended for use in scripts where the script needs to be aware of its own filename,
particularly useful for logging, dynamic referencing, or script self-submission in batch processing environments.
}
\examples{
# To see the name of the script in which this function is called:
script_name <- get_script_name()
print(script_name)
}
