% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SafeExecute.R
\name{handle_error}
\alias{handle_error}
\title{Handle Errors by Printing to Console and Invoking Checkpoint}
\usage{
handle_error(e)
}
\arguments{
\item{e}{The error object captured by tryCatch.}
}
\description{
Prints error messages using a checkpoint function and allows continuation of the script despite errors.
}
\keyword{internal}
