% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDsToArchR.R
\name{IDsToArchR}
\alias{IDsToArchR}
\title{Convert Cell IDs to ArchR Format}
\usage{
IDsToArchR(cell_ids)
}
\arguments{
\item{cell_ids}{A character vector of cell IDs to be converted.}
}
\value{
A character vector of transformed cell IDs in ArchR format.
}
\description{
This function takes a vector of cell IDs and converts them into a specific format
required by ArchR. It handles the prefix transformation and ensures that numeric
parts of the ID are formatted with two leading zeros. This function requires
the \code{\link[stringr]{str_replace}} function from the \pkg{stringr} package.
}
\examples{
IDsToArchR(c("samp_123_cell", "samp_2_cell"))
}
