% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManageSessions.R
\name{init_after_slurm}
\alias{init_after_slurm}
\title{Initialize Environment After Slurm Execution#'}
\usage{
init_after_slurm(session_file_name = NULL)
}
\arguments{
\item{session_file_name}{String; the name of the session file to load. Defaults to the first command line argument.}
}
\description{
This function loads an R session file if it exists and is specifically tailored for use in a Slurm environment.
It aborts if run in an interactive session.
}
\examples{
load_after_slurm()
}
