% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initSmoothR .R
\name{initSmoothR}
\alias{initSmoothR}
\title{Activate runSmoothR.sh by Copying and Adding to $PATH}
\usage{
initSmoothR(force = F)
}
\description{
This function copies 'runSmoothR.sh' to a specified directory and adds the directory
to the $PATH variable by updating the appropriate shell configuration file,
only if the path is not already present.
}
