% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jagged_lines}
\alias{jagged_lines}
\title{Jagged lines for smoothing}
\format{
An \link[sf:sf]{sf} object with 9 features and 3 attributes:
\itemize{
\item \strong{\code{type}}: character; the geometry, i.e. "polygon" or "line".
\item \strong{\code{closed}}: logical; whether the line forms a closed loop or not.
\item \strong{\code{multipart}}: logical; whether the feature is single or multipart.
}
}
\usage{
jagged_lines
}
\description{
Spatial lines in \link[sf:sf]{sf} format for smoothing. There are examples of
lines forming a closed loop and multipart lines.
}
\keyword{datasets}
