% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jagged_lines_3d}
\alias{jagged_lines_3d}
\title{3D jagged line with Z-dimension for smoothing}
\format{
An \link[sf:sf]{sf} object with 9 features and 3 attributes:
\itemize{
\item \strong{\code{type}}: character; the geometry, i.e. "polygon" or "line".
\item \strong{\code{closed}}: logical; whether the line forms a closed loop or not.
\item \strong{\code{multipart}}: logical; whether the feature is single or multipart.
}
}
\usage{
jagged_lines_3d
}
\description{
Spatial lines in \link[sf:sf]{sf}  format for smoothing in three dimensions.
There are examples of open and closed loops
}
\keyword{datasets}
