% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowersect.R
\name{lowersect}
\alias{lowersect}
\title{A case-insensitive intersect.}
\usage{
lowersect(vec1, vec2)
}
\value{
A character vector.
}
\description{
A case-insensitive intersect.
}
\examples{
lowersect(c("A", "B", "C"), c("a", "B", "D"))
}
