% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixEQTLwrapperMC.R
\name{matrixEQTLwrapperMC}
\alias{matrixEQTLwrapperMC}
\title{Run Group eQTL Analysis}
\usage{
matrixEQTLwrapperMC(
  feature_locations_path = NULL,
  feature_data_path = NULL,
  snpFilePath = NULL,
  covFilePath,
  snpLocPath = NULL,
  group_name,
  resultsDir = getwd(),
  cisDist = 1e+06,
  pvOutputThreshold = 1e-05,
  pvOutputThresholdCis = 1e-04,
  useModel = "linear",
  minPvByGeneSnp = TRUE,
  noFDRsaveMemory = FALSE,
  pvalueHist = "qqplot",
  SNPsInChunks = NULL,
  prefix = NULL
)
}
\arguments{
\item{feature_locations_path}{String; path to the file containing genomic feature locations.}

\item{feature_data_path}{String; path to the file containing feature data.}

\item{snpFilePath}{String; path to the file containing SNP data.}

\item{covFilePath}{String; path to the file containing covariate data.}

\item{snpLocPath}{String; path to the file containing SNP locations.}

\item{group_name}{String; name of the group being analyzed.}

\item{resultsDir}{String; directory to store results files, defaults to the current working directory.}

\item{cisDist}{Numeric; maximum distance for considering cis interactions, default is 1e6 (1 million bases).}

\item{pvOutputThreshold}{Numeric; p-value output threshold for trans eQTLs, default is 1e-5.}

\item{pvOutputThresholdCis}{Numeric; p-value output threshold for cis eQTLs, default is 1e-4.}

\item{useModel}{String; statistical model used for the analysis ('linear', 'anova', 'cross'). Default is 'linear'.}

\item{minPvByGeneSnp}{Logical; if TRUE, the minimum p-value by gene-SNP pair will be reported.}

\item{noFDRsaveMemory}{Logical; if TRUE, FDR calculation will not save memory usage.}

\item{pvalueHist}{String; type of p-value histogram to produce ('qqplot', 'histogram', etc.).}

\item{SNPsInChunks}{Logical; indicates if SNP data are in chunks and require specific handling.}

\item{prefix}{Nullable String; additional prefix for output files, helps in identifying files when running multiple models.}
}
\value{
Logical TRUE if the function executes successfully, otherwise it logs an error and stops.
}
\description{
This function performs eQTL analysis for a specified group using SNP and gene expression data across different statistical models. It handles both cis and trans eQTLs and allows for comprehensive output management including saving results to specified paths and optional result plotting.
}
\examples{
matrixEQTLwrapper(
  feature_locations_path = "path/to/features.rds",
  feature_data_path = "path/to/feature_data.rds",
  snpFilePath = "path/to/snps.rds",
  covFilePath = "path/to/covariates.txt",
  snpLocPath = "path/to/snp_locations.txt",
  group_name = "group_cluster1",
  resultsDir = "path/to/results"
)
}
