% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_peak_matrix_by_col.R
\name{normalize_peak_matrix_by_col}
\alias{normalize_peak_matrix_by_col}
\title{Normalize Peak Matrix by Column}
\usage{
normalize_peak_matrix_by_col(x, norm_col_name, test = NULL, scale = FALSE, ...)
}
\arguments{
\item{x}{A `SummarizedExperiment` object.}

\item{norm_col_name}{The name of the column in `colData` used for normalization.}

\item{test}{Optional integer to subset the matrix for testing.}
}
\value{
Returns a normalized matrix, or a list of matrices if testing.
}
\description{
This function normalizes a peak matrix within a SummarizedExperiment object
by a specified column from `colData`.
}
\examples{
seRNA <- SummarizedExperiment::SummarizedExperiment(assays = list(PeakMatrix = matrix))
normalize_peak_matrix_by_col(seRNA, "norm_factor")
}
