% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_args.R
\name{print_args}
\alias{print_args}
\title{Print Command Line Arguments}
\usage{
print_args()
}
\value{
Prints each command-line argument passed to the script; returns invisible NULL in interactive mode.
}
\description{
This function prints all command line arguments provided to the R script
when run from the command line. It will not execute in interactive sessions
to ensure it is used specifically for command-line operations.
}
