% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManageSessions.R
\name{print_checkpoint}
\alias{print_checkpoint}
\title{Print checkpoint messages for Slurm job monitoring}
\usage{
print_checkpoint(phrase)
}
\arguments{
\item{phrase}{A character string representing the message to be printed.}
}
\description{
This function prints a specified message with a 'CHECKPOINT_' prefix, which is used to signal specific stages
or statuses in a Slurm job script. This is particularly useful for logging and monitoring job progress.
}
\examples{
print_checkpoint("data_loaded")
}
