% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManageSessions.R
\name{quit_success}
\alias{quit_success}
\title{Quit R session}
\usage{
quit_success()
}
\description{
This function terminates the current R session and exits with status 0, indicating successful completion.
It is particularly useful for scripting and batch processing in environments like Slurm.
}
\examples{
quit_success()
}
