% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_slurm.R
\name{run_slurm}
\alias{run_slurm}
\title{Submit the Current R Script to Slurm for Execution}
\usage{
run_slurm(output_file = NULL)
}
\arguments{
\item{output_file}{An optional string specifying the file path where the Slurm job output should be redirected.
If not specified, the output will be handled according to the Slurm script's configuration.}
}
\value{
The function does not return a value but prints the submission status to the console.
        It is designed for interactive use to allow dynamic job control.
}
\description{
This function automatically submits the currently running R script to a Slurm-managed cluster
using a predefined Slurm script named 'run_loop.sh'. The function assumes that the Slurm script
is available in the system's PATH and is properly configured to accept an R script as a parameter.
The function can optionally redirect the output of the Slurm job to a specified file.
}
\examples{
# Submit the current script to Slurm, outputting to the console
run_slurm()

# Submit the current script to Slurm, redirecting output to 'job_output.txt'
run_slurm("job_output.txt")
}
