% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExpression.R
\name{runExpression}
\alias{runExpression}
\title{Run an R expression and submit it to Slurm for execution.}
\usage{
runExpression(r_expression, save = FALSE)
}
\arguments{
\item{r_expression}{The R expression to run.}
}
\value{
NA
}
\description{
This function runs an R expression and submits it to Slurm for execution.
}
