% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runScript.R
\name{runScript}
\alias{runScript}
\title{Run a R script and submit it to Slurm for execution.}
\usage{
runScript(script_name, wait = FALSE, lint = FALSE)
}
\arguments{
\item{script_name}{The name of the R script to run.}

\item{lint}{Whether to lint the script instead of running it.}

\item{output_file}{The name of the output file.}
}
\value{
NA
}
\description{
This function runs a R script and submits it to Slurm for execution.
}
