% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSlurm.R
\name{RunSlurm}
\alias{RunSlurm}
\title{Submit the Running R Script to Slurm for Execution}
\usage{
RunSlurm(output_file = NULL)
}
\arguments{
\item{output_file}{Optional; specifies the file path where the Slurm job output should be redirected.
If not specified, the output will be displayed in the console.}
}
\value{
The function does not return a value but prints the submission status to the console.
        It is designed for interactive use to allow dynamic job control.
}
\description{
This function submits the currently running R script to a Slurm-managed cluster.
It uses a predefined Slurm script (`run_R_smootly.sh`) assumed to be available in the system's PATH.
The function automatically retrieves the name of the running script and submits it to Slurm,
which makes the script handling secure and error-free. Also, a function (`activate()` will copy the file to your folder of choice (profided ar argument), and add the route to it to your $PATH variable by updatin "~/.bashrc", or ~/.zshrc
}
\examples{
# Submit the currently running script to Slurm, outputting to the console
RunSlurm()

# Submit the current script to Slurm, redirecting output to 'job_output.txt'
RunSlurm("job_output.txt")
}
