% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SafeExecute.R
\name{SafeExecute}
\alias{SafeExecute}
\title{General Purpose Try-Catch Wrapper}
\usage{
SafeExecute(expr)
}
\arguments{
\item{expr}{An expression to evaluate.}
}
\description{
Executes code with robust error and warning handling to prevent script termination.
The function utilizes a custom checkpoint system to manage and log all issues.
}
\keyword{internal}
