% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManageSessions.R
\name{save_for_slurm}
\alias{save_for_slurm}
\title{Save Current R Session for Slurm Execution}
\usage{
save_for_slurm(session_file_name = NULL)
}
\arguments{
\item{session_file_name}{String; the filename where the session should be saved. Defaults to the first command line argument.}
}
\description{
Saves the current R session into a file, allowing for future continuation in Slurm job executions.
It aborts if run in an interactive session.
}
\examples{
save_for_slurm()
}
