% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveNow.R
\name{SaveNow}
\alias{SaveNow}
\title{Save the Current R Session for Slurm Execution and Log Session Information}
\usage{
SaveNow(
  session_file_name = NULL,
  timestamp = format(Sys.time(), "\%Y\%m\%d-\%H\%M\%S")
)
}
\arguments{
\item{session_file_name}{string (optional); the filename for saving the session.
If not provided, the first command line argument will be used with "Data" appended.}

\item{timestamp}{string (optional); the timestamp to append to the session info file.
Defaults to the current datetime if not provided.}
}
\description{
This function saves the current R session into a .RData file, enabling the session
to be resumed in future Slurm job executions. It also logs session details
into a timestamped text file, aiding in reproducibility and debugging.
The session file name and timestamp can be specified; if not, defaults are used.
}
\examples{
SaveNow()  # Save using default settings
}
