% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothr-package.R
\docType{package}
\name{smoothr}
\alias{smoothr}
\title{\code{smoothr}: Smooth and Tidy Spatial Features}
\description{
Tools for smoothing and tidying spatial features (i.e. lines and polygons) to
make them more aesthetically pleasing. Smooth curves, fill holes, and remove
small fragments from lines and polygons.
}
