% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetSE.R
\name{subsetSE}
\alias{subsetSE}
\title{Subset a SummarizedExperiment Object}
\usage{
subsetSE(SE, nrow = NULL, ncol = NULL, byCol = NULL)
}
\arguments{
\item{SE}{A SummarizedExperiment object.}

\item{nrow}{Optional; the number of rows to randomly sample from the SE object.
If more than the available rows, all rows are used.}

\item{ncol}{Optional; the number of columns to randomly sample from the SE object.
If more than the available columns, all columns are used.}

\item{byCol}{Optional; a vector where the first element is the name of the column
in 'colData' used for subsetting, followed by values to filter by.
Only character subsetting is supported.}
}
\value{
A subsetted SummarizedExperiment object.
}
\description{
This function subsets a SummarizedExperiment object based on metadata criteria
and/or random sampling of rows and columns.
}
\examples{
\dontrun{
  # Assuming `se` is your SummarizedExperiment object
  subset_se <- subsetSE(se, nrow = 100, ncol = 50, byCol = c("condition", "treated", "control"))
}
}
