% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_peak_matrix.R
\name{summarize_peak_matrix}
\alias{summarize_peak_matrix}
\title{Summarizes Peak Matrix Scores by Cell from Sparse Matrices in Parallel}
\usage{
summarize_peak_matrix(sparse_matrix, colData, colNames)
}
\arguments{
\item{sparse_matrix}{A sparse matrix object from the Matrix package, expected to be
the output of a peak matrix computation.}

\item{colData}{A data.table or convertible structure containing metadata for the matrix
columns. This parameter must include all columns specified in 'colNames'.}

\item{colNames}{A character vector specifying the names of columns in 'colData' that
will be used for grouping and parallel computation.}
}
\value{
A data.table containing the summarized occurrence counts of matrix values, grouped
        by the specified 'colNames' and including additional identifiers and group information.
}
\description{
This function processes a sparse matrix to summarize peak scores for each cell,
based on a peak matrix generated by the Archer getPeakMatrix function.
It efficiently handles large datasets using data.table for data manipulation
and parallel computing to enhance performance. It requires 'colData' with specific
columns as indicated in 'colNames'.
}
\examples{
# Create a sparse matrix
library(Matrix)
sparse_matrix <- Matrix(c(0,0,2:0), 3, 5, sparse = TRUE)

# Create column data as a data.table
library(data.table)
colData <- data.table(cell_names = c("Cell1", "Cell2", "Cell3"),
                      group = c("A", "A", "B"))
colNames <- c("group")

# Calculate occurrences
results <- summarize_peak_matrix(sparse_matrix, colData, colNames)
print(results)

}
