% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{update_num_threads}
\alias{update_num_threads}
\title{Update the Number of Threads}
\usage{
update_num_threads(NUM_THREADS)
}
\arguments{
\item{NUM_THREADS}{New number of threads to set.}
}
\description{
Update the Number of Threads
}
