% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{update_output_dir}
\alias{update_output_dir}
\title{Update Output Directory}
\usage{
update_output_dir(OUTPUT_DIR)
}
\arguments{
\item{OUTPUT_DIR}{New output directory to set.}
}
\description{
Update Output Directory
}
