% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{UpdateConfig}
\alias{UpdateConfig}
\title{Update Configuration Setting}
\usage{
UpdateConfig(key, value)
}
\arguments{
\item{key}{The configuration key to update (character).}

\item{value}{The new value for the key (character).}
}
\value{
Invisible NULL; updates the file in-place.
}
\description{
This function updates a specific configuration key in the `.temp_shell_exports`
file located in the user's home directory. If the key does not exist, it adds the key with the specified value.
}
\examples{
UpdateConfig("NUM_THREADS", "8") # This example updates the number of threads.
}
