% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{UpdateJobTime}
\alias{UpdateJobTime}
\title{Update Job Time}
\usage{
UpdateJobTime(JOB_TIME_MINUTES)
}
\arguments{
\item{JOB_TIME_MINUTES}{New job time to set in minutes.}
}
\description{
Update Job Time
}
