% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{updateNumThreads}
\alias{updateNumThreads}
\title{Update the Number of Threads}
\usage{
updateNumThreads(NUM_THREADS)
}
\arguments{
\item{NUM_THREADS}{New number of threads to set.}
}
\description{
Update the Number of Threads
}
