% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{updateOutputDir}
\alias{updateOutputDir}
\title{Update Output Directory}
\usage{
updateOutputDir(OUTPUT_DIR)
}
\arguments{
\item{OUTPUT_DIR}{New output directory to set.}
}
\description{
Update Output Directory
}
