% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{AddFat}
\alias{AddFat}
\title{Add or Update FAT Configuration}
\usage{
AddFat(fat_value = NULL)
}
\arguments{
\item{fat_value}{Optional character indicating the FAT setting ('T', 'F', 't', 'f'); default NULL.}
}
\description{
Adds or updates the FAT setting in the configuration file. It validates and accepts
only 'T', 'F', 't', 'f' as values. If no value is provided, the current configuration is printed.
}
