% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{checkDir}
\alias{checkDir}
\title{Change Working Directory or Verify Current Directory}
\usage{
checkDir(output_dir)
}
\arguments{
\item{output_dir}{The target directory to validate or change to.}
}
\description{
Verifies if the specified `output_dir` exists and changes the working directory to it.
Throws an error if the directory does not exist or does not match the current directory.
}
