% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManageSessions.R
\name{InitNow}
\alias{InitNow}
\title{Initialize Environment After Slurm Execution}
\usage{
InitNow(session_file_name = NULL)
}
\arguments{
\item{session_file_name}{String; the name of the session file to load.
Defaults to the first command line argument.}
}
\description{
This function loads an R session file if it exists.
It aborts if run in an interactive session. If the specified session file does not exist, the function will continue
without loading it, ensuring that the show goes on.
}
\examples{
InitNow()
}
