% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageSessions.R
\name{InitNow}
\alias{InitNow}
\title{Initialize Session After Slurm Execution}
\usage{
InitNow(session_file_name = NULL)
}
\arguments{
\item{session_file_name}{The name of the session file to load, defaults to the first command line argument with ".RData" appended.}
}
\description{
Loads a specified R session file. Aborts if run interactively or if the session file does not exist.
Continues without loading if the file is missing.
}
\examples{
InitNow()
}
