% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitSmoothR .R
\name{InitSmoothR}
\alias{InitSmoothR}
\title{Activate runSmoothR.sh by Copying and Adding to $PATH}
\usage{
InitSmoothR()
}
\description{
This function copies 'runSmoothR.sh' to a specified directory and adds the directory
to the $PATH variable by updating the appropriate shell configuration file,
only if the path is not already present.
}
