% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jagged_raster}
\alias{jagged_raster}
\title{Simulated raster for polygonizing and smoothing}
\format{
An object of class \code{RasterLayer} of dimension 25 x 25 x 1.
}
\usage{
jagged_raster
}
\description{
One of the primary applications of this package is for smoothing polygons
generated from rasters. This example RasterLayer is meant to be a simulated
occurrence probability for a species, consisting of a spatially
auto-correlated Gaussian field with values between 0 and 1. This raster is a
25x25 grid of 100 square kilometer cells in a North American centered Alber's
Equal Area projection.
}
\keyword{datasets}
