% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManageSessions.R
\name{QuitNow}
\alias{QuitNow}
\title{Quit R Session when not Interactive}
\usage{
QuitNow()
}
\description{
Terminates the current R session non-interactively with a success status. Useful in scripts and batch processing.
Does not save the session before exiting. Precede with SaveNow() to save data first.
}
\examples{
QuitNow()
}
