% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ReadFromConfig}
\alias{ReadFromConfig}
\title{Read Configuration Values from a File}
\usage{
ReadFromConfig(key_to_check)
}
\arguments{
\item{key_to_check}{The configuration key to retrieve.}

\item{config_file}{Path to the configuration file, defaults to '~/.temp_shell_exports'.}
}
\value{
The value associated with the provided key.
}
\description{
Retrieves a specific configuration value by key from a file formatted as `export KEY='VALUE'`.
Throws an error if the key is not found or if the file does not exist.
}
