% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunNow.R
\name{RunNow}
\alias{RunNow}
\title{Run a R script and submit it to Slurm for execution.}
\usage{
RunNow(script_name, output_file = NULL, lint = FALSE, verbose = FALSE)
}
\arguments{
\item{script_name}{The name of the R script to run.}

\item{output_file}{The name of the output file.}

\item{lint}{Whether to lint the script instead of running it.}

\item{verbose}{Whether to print verbose output.}
}
\value{
NA
}
\description{
This function runs a R script and submits it to Slurm for execution.
}
