% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunNow.R
\name{RunNow}
\alias{RunNow}
\title{RunNow Function}
\usage{
RunNow()
}
\arguments{
\item{config_file}{A string specifying the path to the configuration file that sets required
environment variables. Defaults to "~/.temp_shell_exports".}

\item{output_file}{A string specifying the path to the file where stdout and stderr from the Slurm
job will be directed. If NULL, outputs will not be redirected. Defaults to NULL.}
}
\value{
The value of the `OUTPUT_DIR` environment variable, indicating the output directory.
}
\description{
This function is designed to submit R scripts to a Slurm job scheduler. It reads configuration from
a specified shell script in the user's home directory to set environment variables in R, retrieves 
the `OUTPUT_DIR` environment variable, and submits the job using a specified Slurm script.
}
\examples{
RunNow()  # Uses default configuration file and no output redirection
RunNow(output_file = "path/to/output.log")
}
