% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{setAccount}
\alias{setAccount}
\title{Set COMPUTE_ACCOUNT Environment Variable}
\usage{
setAccount(accountName)
}
\arguments{
\item{accountName}{Account name to set as COMPUTE_ACCOUNT.}
}
\description{
Sets the COMPUTE_ACCOUNT environment variable to the specified account name and displays it.
}
\examples{
setAccount("snic123456")
}
