% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setConfig.R
\name{setConfig}
\alias{setConfig}
\title{Set Environment Configuration Interactively}
\usage{
setConfig(force = F)
}
\arguments{
\item{None}{Parameters are input interactively.}
}
\value{
No return value; outputs file path and contents to console.
}
\description{
Interactively prompts for various configuration settings and writes them to a shell script in the user's home directory.
The script, formatted for environment variable export, is intended for shell sourcing.
Includes conversion of job time from minutes to D-H:M:S format.
}
\examples{
setConfig() # Run this in an interactive R session
}
