% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetConfig.R
\name{SetConfig}
\alias{SetConfig}
\title{Set Environment Configuration Interactively}
\usage{
SetConfig(force = F)
}
\arguments{
\item{None}{Parameters are input interactively.}
}
\value{
No return value; outputs file path and contents to console.
}
\description{
Interactively prompts for various configuration settings and writes them to a shell script in the user's home directory.
The script, formatted for environment variable export, is intended for shell sourcing.
Includes conversion of job time from minutes to D-H:M:S format.
}
\examples{
set_config_interactively() # Run this in an interactive R session
}
