% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetModules.R
\name{SetModules}
\alias{SetModules}
\title{Update or Create Module Configuration}
\usage{
SetModules(modules = c("R/4.1.1", "R_packages/4.1.1"), force = TRUE)
}
\arguments{
\item{modules}{A character vector of module names to be loaded (default is c("R/4.1.1", "R_packages/4.1.1")).}

\item{force}{Logical; if TRUE, forcefully overwrite the existing module file if it exists (default is TRUE).}
}
\value{
No return value; outputs file path and contents to console.
}
\description{
Updates or creates a module configuration file in a specified directory, overwriting the existing file if allowed. The function is designed for interactive R sessions only and prints the file path and content to the console.
}
\examples{
SetModules() # Default usage
SetModules(modules = c("Python/3.8.5", "Java/1.8"), force = FALSE)
}
