% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.r
\name{smooth}
\alias{smooth}
\title{Smooth a spatial feature}
\usage{
smooth(x, method = c("spline"))
}
\arguments{
\item{x}{spatial features; lines or polygons from either the sf or sp
packages.}

\item{method}{character; specifies the type of smoothing method to use.
Possible methods are:
\itemize{
\item \code{"spline"}: uses Bézier spline interpolation via the \code{\link[stats:spline]{stats::spline()}}
function. This method interpolates between existing vertices and the
resulting smoothed feature will pass through the vertices of the input
feature.
}}
}
\value{
A smoothed polygon or line in the same format as the input data.
}
\description{
Smooth a spatial feature
}
\examples{
library(sf)
# spline smoothing
# example polygons
po <- par(mar = c(0, 0, 0, 0), oma = c(0, 0, 2, 0), mfrow = c(3, 3))
for (i in 1:nrow(jagged_polygons)) {
  p <- jagged_polygons[i, ]
  smoothed <- smooth(p, method = "spline")
  plot(st_geometry(smoothed), col = NA, border = NA)
  plot(st_geometry(p), col = "grey20", border = NA, add = TRUE)
  plot(st_geometry(smoothed), col = NA, border = "red", lwd = 2, add = TRUE)
  title("Smoothed Polygons (Bézier Spline)", cex.main = 2, outer = TRUE)
}
par(po)

# example lines
po <- par(mar = c(0, 0, 0, 0), oma = c(0, 0, 2, 0), mfrow = c(3, 3))
for (i in 1:nrow(jagged_lines)) {
  l <- jagged_lines[i, ]
  smoothed <- smooth(l, method = "spline")
  plot(st_geometry(smoothed), col = NA)
  plot(st_geometry(l), col = "grey20", lwd = 2, add = TRUE)
  plot(st_geometry(smoothed), col = "red", lwd = 2, add = TRUE)
  title("Smoothed Lines (Bézier Spline)", cex.main = 2, outer = TRUE)
}
par(po)
}
\references{
This function was inspired by the following StackExchange posts:
\itemize{
\item \href{https://stackoverflow.com/questions/26087772/#26089377}{Create polygon from set of points distributed}
\item \href{https://gis.stackexchange.com/questions/24827/#24929}{Smoothing polygons in contour map?}
}
}
