% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth-chaikin.r
\name{smooth_chaikin}
\alias{smooth_chaikin}
\title{Chaikin's corner cutting algorithm}
\usage{
smooth_chaikin(x, type = c("polygon", "line"), refinements = 4L)
}
\arguments{
\item{x}{numeric matrix; 2-column matrix of coordinates.}

\item{type}{character; whether the coordinates correspond to a line or
polygon.}

\item{refinements}{integer; number of corner cutting iterations to apply.}
}
\value{
A matrix with the coordiantes of the smoothed curve.
}
\description{
Chaikin's corner cutting algorithm smooths a curve by iteratively replacing
every point by two new points: one 1/4 of the way to the next point and one
1/4 of the way to the previous point. This method results in a curve that
does not pass through the vertices of the original curve.
}
\details{
This function works on matrices of points and is generally not called
directly. Instead, use \code{\link[=smooth]{smooth()}} with \code{method = "chaikin"} to apply this
smoothing algorithm to spatial features.
}
\examples{
library(sf)
# smooth polygons
p_smoothed <- smooth(jagged_polygons, method = "chaikin")
par(mar = c(0, 0, 0, 0), oma = c(0, 0, 2, 0), mfrow = c(3, 3))
for (i in 1:nrow(jagged_polygons)) {
  plot(st_geometry(jagged_polygons[i, ]), col = "grey20", border = NA)
  plot(st_geometry(p_smoothed[i, ]), col = NA, border = "red", lwd = 2, add = TRUE)
  title("Smoothed Polygons (Chaikin's Corner Cutting)", cex.main = 2, outer = TRUE)
}

# smooth lines
l_smoothed <- smooth(jagged_lines, method = "chaikin")
par(mar = c(0, 0, 0, 0), oma = c(0, 0, 2, 0), mfrow = c(3, 3))
for (i in 1:nrow(jagged_lines)) {
  plot(st_geometry(jagged_lines[i, ]), col = "grey20", lwd = 2)
  plot(st_geometry(l_smoothed[i, ]), col = "red", lwd = 2, add = TRUE)
  title("Smoothed Lines (Chaikin's Corner Cutting)", cex.main = 2, outer = TRUE)
}
}
\references{
The orginal reference for Chaikin's corner curring algorithm is:
\itemize{
\item \code{Chaikin, G. An algorithm for high speed curve generation. Computer Graphics and Image Processing 3 (1974), 346–349}
}

This implementation was inspired by the following StackOverflow answer:
\itemize{
\item \href{https://stackoverflow.com/a/47255374/3591386}{Where to find Python implementation of Chaikin's corner cutting algorithm?}
}
}
\seealso{
\code{\link[=smooth]{smooth()}}
}
