% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth-spline.r
\name{smooth_spline}
\alias{smooth_spline}
\title{Spline interpolation}
\usage{
smooth_spline(x, type = c("polygon", "line"), n = 100, smoothness = NULL)
}
\arguments{
\item{x}{numeric matrix; 2-column matrix of coordinates.}

\item{type}{character; whether the coordinates correspond to a line or
polygon.}

\item{n}{integer; number of vertices in the smoothed curve. Ignored if
\code{smoothness} is specified.}

\item{smoothness}{double; the proportional increase in the number of vertices
in the smooth curve. For example, if the oringal curve has 10 points, a
value of \code{2.5} will yield a new smoothed curve with 250 points.}
}
\value{
A matrix with the coordiantes of the smoothed curve.
}
\description{
Spline interpolation uses \code{\link[stats:spline]{stats::spline()}} to interpolate between existing
vertices using piecewise cubic polynomials. The \code{x} and \code{y} coordinates are
interpolated independently, and .
}
\details{
This function works on matrices of points and is generally not called
directly. Instead, use \code{\link[=smooth]{smooth()}} with \code{method = "spline"} to apply this
smoothing algorithm to spatial features.
}
\examples{
library(sf)
# smooth polygons
p_smoothed <- smooth(jagged_polygons, method = "spline")
par(mar = c(0, 0, 0, 0), oma = c(0, 0, 2, 0), mfrow = c(3, 3))
for (i in 1:nrow(jagged_polygons)) {
  plot(st_geometry(p_smoothed[i, ]), col = NA, border = NA)
  plot(st_geometry(jagged_polygons[i, ]), col = "grey20", border = NA, add = TRUE)
  plot(st_geometry(p_smoothed[i, ]), col = NA, border = "red", lwd = 2, add = TRUE)
  title("Smoothed Polygons (Spline Interpolation)", cex.main = 2, outer = TRUE)
}

# smooth lines
l_smoothed <- smooth(jagged_lines, method = "spline")
par(mar = c(0, 0, 0, 0), oma = c(0, 0, 2, 0), mfrow = c(3, 3))
for (i in 1:nrow(jagged_lines)) {
  plot(st_geometry(l_smoothed[i, ]), col = NA)
  plot(st_geometry(jagged_lines[i, ]), col = "grey20", lwd = 2, add = TRUE)
  plot(st_geometry(l_smoothed[i, ]), col = "red", lwd = 2, add = TRUE)
  title("Smoothed Lines (Spline Interpolation)", cex.main = 2, outer = TRUE)
}
}
\references{
The spline method was inspired by the following StackExchange
posts:
\itemize{
\item \href{https://stackoverflow.com/questions/26087772/26089377}{Create polygon from set of points distributed}
\item \href{https://gis.stackexchange.com/questions/24827/24929}{Smoothing polygons in contour map?}
}
}
\seealso{
\code{\link[=smooth]{smooth()}}
}
