% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothr-package.R
\docType{package}
\name{smoothr-package}
\alias{smoothr}
\alias{smoothr-package}
\title{\code{smoothr}: Smooth and Tidy Spatial Features}
\description{
Tools for smoothing and tidying spatial features (i.e. lines and polygons) to
make them more aesthetically pleasing. Smooth curves, fill holes, and remove
small fragments from lines and polygons.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://strimas.com/smoothr/}
  \item \url{https://github.com/mstrimas/smoothr}
  \item Report bugs at \url{https://github.com/mstrimas/smoothr/issues}
}

}
\author{
\strong{Maintainer}: Matthew Strimas-Mackey \email{mstrimas@gmail.com} (\href{https://orcid.org/0000-0001-8929-7776}{ORCID})

}
\keyword{internal}
