% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{updateConfig}
\alias{updateConfig}
\title{Update Configuration Setting}
\usage{
updateConfig(key, value)
}
\arguments{
\item{key}{The configuration key to update.}

\item{value}{The new value for the key.}
}
\value{
Invisible NULL; updates are made in-place.
}
\description{
This function updates a specific configuration key in the `.temp_shell_exports`
file located in the user's home directory. If the key does not exist, it adds the key with the specified value.
}
\examples{
updateConfig("NUM_THREADS", "8")
}
