% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{updateJobTime}
\alias{updateJobTime}
\title{Update Job Time}
\usage{
updateJobTime(JOB_TIME_MINUTES)
}
\arguments{
\item{JOB_TIME_MINUTES}{New job time to set in minutes.}
}
\description{
Update Job Time
}
