% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{UpdateNumThreads}
\alias{UpdateNumThreads}
\title{Update the Number of Threads}
\usage{
UpdateNumThreads(NUM_THREADS)
}
\arguments{
\item{NUM_THREADS}{New number of threads to set.}
}
\description{
Update the Number of Threads
}
